+++ client/components/icons/CrossIcon.tsx
