+++ client/components/icons/CrossIconWhite.tsx
