+++ client/components/icons/ExternalLinkIcon.tsx
