+++ client/components/layout.tsx
@@ -1,7 +1,7 @@
 import { FunctionComponent, ReactNode } from "react";
 import Header from "components/Header";
 import Footer from "components/Footer";
-import AdminUtils from "components/AdminUtils";
+import AdminUtils from "components/utils/AdminUtils";
 
 interface LayoutProps {
   children: ReactNode;
