+++ client/utils/useHistoricalLockupToasts.js
@@ -7,7 +7,7 @@ import { truncateEthAddress } from "utils";
 import { SECONDS_IN_A_MONTH } from "../constants/index";
 import TokenAmount from "components/TokenAmount";
 import Link from "components/Link";
-import ExternalLinkIcon from "components/ExternalLinkIcon";
+import ExternalLinkIcon from "components/icons/ExternalLinkIcon";
 
 const _LockupContent = ({ shortAddress, ogvLockedUp, months, txHash }) => {
   return (
